Load interface textures.

+++ widgets/eprogressbar.cpp
@@ -21,7 +21,7 @@ void eProgressBar::sizeHint(int& w, int& h) {
 
 void eProgressBar::paintEvent(ePainter& p) {
     p.drawRect(rect(), {0, 0, 0, 255}, 2);
-    const double per = (mValue - mMin)/(mMax - mMin);
+    const double per = double(mValue - mMin)/(mMax - mMin);
     const int pa = padding();
     const int w = per*(width() - 2*pa);
     const SDL_Rect rect{pa, pa, w, height() - 2*pa};
